Modificacion de diseño
nuevo orden de diseño
