+++ proyectodps2022/Navigation.js
@@ -3,15 +3,61 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { NavigationContainer } from '@react-navigation/native';
 import { MaterialCommunityIcons } from '@expo/vector-icons'; 
 
-//screens
+//pantallas de autentificacion
+import LoginScreen from './authscreens/LoginScreen';
+import RegisterScreen from './authscreens/RegisterScreen';
+
+//pantallas dentro de la navegacion de la app
 import DietScreen from './screens/DietScreen';
-import LoginScreen from './screens/LoginScreen';
 import ProfileScreen from './screens/ProfileScreen';
-import RegisterScreen from './screens/RegisterScreen';
 import RoutineScreen from './screens/RoutineScreen';
 import HomeScreen from './screens/HomeScreen';
 
 const Tab = createBottomTabNavigator();
+
+function AuthTabs() {
+  return (
+    <Tab.Navigator
+    initialRouteName="Sesion"
+    screenOptions={{
+      tabBarActiveTintColor : 'blue',
+      tabBarShowLabel : false,
+      tabBarStyle : {
+        height : 70,
+        position : 'absolute',
+        bottom : 16,
+        right : 16,
+        left : 16,
+        borderRadius : 16,
+        showLabel : false,
+      },
+    }}
+    >
+     <Tab.Screen
+        name="Sesion"
+        component={LoginScreen}
+        options={{
+          //tabBarLabel: 'Iniciar Sesion',
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons name="account-cog" size={24} color="black" />
+          ),
+          headerShown: false,
+        }}
+        />
+     <Tab.Screen
+        name="Registrarse"
+        component={RegisterScreen}
+        options={{
+          //tabBarLabel: 'Registrarse',
+          tabBarIcon: ({ color, size }) => (
+            <MaterialCommunityIcons name="brightness-5" size={24} color="black" />
+          ),
+          headerShown: false,
+        }}
+        />
+    </Tab.Navigator>
+  );
+}
 
 function MyTabs() {
   return (
@@ -19,13 +65,23 @@ function MyTabs() {
     initialRouteName="Inicio"
     screenOptions={{
       tabBarActiveTintColor : 'blue',
+      tabBarShowLabel : false,
+      tabBarStyle : {
+        height : 70,
+        position : 'absolute',
+        bottom : 16,
+        right : 16,
+        left : 16,
+        borderRadius : 16,
+        showLabel : false,
+      },
     }}
     >
     <Tab.Screen
         name="Inicio"
         component={HomeScreen}
         options={{
-          tabBarLabel: 'Inicio',
+          //tabBarLabel : 'Inicio',
           tabBarIcon: ({ color, size }) => (
             <MaterialCommunityIcons name="home" size={24} color="black" />
           ),
@@ -36,7 +92,7 @@ function MyTabs() {
         name="Perfil"
         component={ProfileScreen}
         options={{
-          tabBarLabel: 'Perfil',
+          //tabBarLabel: 'Perfil',
           tabBarIcon: ({ color, size }) => (
             <MaterialCommunityIcons name="account" size={24} color="black" />
           ),
@@ -47,7 +103,7 @@ function MyTabs() {
         name="Dietas"
         component={DietScreen}
         options={{
-          tabBarLabel: 'Dietas',
+          //tabBarLabel: 'Dietas',
           tabBarIcon: ({ color, size }) => (
             <MaterialCommunityIcons name="nutrition" size={24} color="black" />
           ),
@@ -58,31 +114,9 @@ function MyTabs() {
         name="Rutinas"
         component={RoutineScreen}
         options={{
-          tabBarLabel: 'Rutinas',
+          //tabBarLabel: 'Rutinas',
           tabBarIcon: ({ color, size }) => (
             <MaterialCommunityIcons name="weight-lifter" size={24} color="black" />
-          ),
-          headerShown: false,
-        }}
-        />
-     <Tab.Screen
-        name="Sesion"
-        component={LoginScreen}
-        options={{
-          tabBarLabel: 'Iniciar Sesion',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name="account-cog" size={24} color="black" />
-          ),
-          headerShown: false,
-        }}
-        />
-     <Tab.Screen
-        name="Registro"
-        component={RegisterScreen}
-        options={{
-          tabBarLabel: 'Registrarse',
-          tabBarIcon: ({ color, size }) => (
-            <MaterialCommunityIcons name="brightness-5" size={24} color="black" />
           ),
           headerShown: false,
         }}
